backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
click==8.2.1
fastapi==0.129.0
fsspec==2026.2.0
numpy==2.2.6
opencv-python==4.12.0.88
pillow==11.3.0
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.2.1
requests==2.32.5
starlette==0.52.1
torch==2.10.0
torchvision==0.25.0
typing_extensions==4.15.0
uvicorn==0.34.0
python-multipart==0.0.9
google-generativeai==0.8.4
tqdm==4.67.3
